+++ src/common/ui/Button/Button.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf, action } from '@storybook/react';
 
-import { Button } from 'components';
+import { Button } from 'ui';
 
 storiesOf('Global/Button', module)
     .add('default', () => (
