+++ src/common/ui/Input/Input.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import { Input } from 'components';
+import { Input } from 'ui';
 
 storiesOf('Global/Input', module)
     .add('default', () => <Input />)
