+++ src/common/ui/Loader/Loader.scss
