+++ src/common/ui/SvgIcon/SvgIcon.jsx
