+++ src/common/ui/SvgIcon/SvgIcon.scss
