+++ src/common/ui/SvgIcon/icons.js
