+++ src/common/ui/Text/Text.scss
