+++ src/components/App/Home/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from 'components';
+import { Text } from 'ui';
 
 import styles from './Home.scss';
