+++ src/components/App/NotFound/NotFound.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Text } from 'components';
+import { Text } from 'ui';
 
 export default function NotFound() {
     return (
